+++ R/get-rho.R
@@ -1,0 +1,37 @@
+#' @title MLE of \eqn{\rho}
+#' 
+#' @description MLE of \eqn{\rho}.
+#' 
+#' @param x Covariate matrix.
+#' @param y Response vector.
+#' @param w Weight matrix (row-sum scaled).
+#' 
+#' @return MLE of \eqn{\rho}.
+#' 
+#' @examples
+#' set.seed(2025)
+#' b0 <- c(1.5, 3.0, 2.0, rep(0.0, 3))
+#' rho0 <- 0.2
+#' sig0 <- 1.0
+#' n <- 81
+#' 
+#' DF <- simu_sam_data_rook(b0, rho0, sig0, n)
+#' system.time( get_rho(as.matrix(DF[["X"]]), DF[["y"]], DF[["W0"]]) )
+#' 
+#' @export
+get_rho <- function(x, y, w) {
+    stopifnot( NROW(x) == length(y) )
+    stopifnot( all(rowSums(w) == 1) )
+    stopifnot( is.matrix(x) )
+
+    if (NCOL(x) > NROW(x))
+        warning("Singularity Possible!")
+
+    minusloglik <- function(rho) {
+        A.rho <- diag(NROW(x)) - rho * w
+        sig2.hat <- mean( residuals(lm.fit(x, A.rho %*% y))^2 ) # singular.ok=FALSE
+        0.5*NROW(x) * log(sig2.hat) - log(det(A.rho))
+    }
+
+    optimize(minusloglik, c(-1, 1))$minimum
+}