+++ R/plagsarlm.R
@@ -1,8 +1,11 @@
-#' @name pboostsam
+#' @name plagsarlm
 #' 
 #' @title Profile Boosting Feature Selection for Spatial Autoregressive Model
 #' @description Profile Boosting Feature Selection for Spatial Autoregressive Model
 #' 
+#' @param x Matrix of covariates.
+#' @param y Vector of response.
+#' @param w Weight matrix.
 #' @param formula Parameters passed to [spatialreg::lagsarlm].
 #' @param data Parameters passed to [spatialreg::lagsarlm].
 #' @param listw Parameters passed to [spatialreg::lagsarlm].
@@ -19,8 +22,8 @@
 #' @param stopFun Stopping rule for profile boosting, which has the form
 #'  stopFun(object) to evaluate the performance of model object returned
 #'  by fitFun, such as [EBIC] or [BIC].
-#' @param maxK Maximal number of identified features. 
-#' If `maxK` is specified, it will supress `stopFun`, saying that the 
+#' @param maxK Maximal number of identified features.
+#' If `maxK` is specified, it will supress `stopFun`, saying that the
 #' profile boosting continues until the procedure identifies `maxK` features.
 #' The pre-specified features in `keep` are counted toward `maxK`.
 #' @param keep Initial set of features that are included in model fitting.
@@ -42,8 +45,10 @@
 #' DF <- simu_sam_data_rook(b0, rho0, sig0, n)
 #' 
 #' data <- with(DF, data.frame(y, X))
-#' W0 <- DF[["W0"]]
-#' ( result <- pboostsam(y ~ ., data, mat2listw(W0, style="W"), verbose=TRUE) )
+#' W0 <- mat2listw(DF[["W0"]], style="W")
+#' system.time( egg1 <- plagsarlm(y ~ ., data, W0, verbose=TRUE) )
+#' 
+#' system.time( egg2 <- plagsarlm2(as.matrix(DF[["X"]]), DF[["y"]], DF[["W0"]], verbose=TRUE) )
 #' 
 #' 
 NULL
@@ -51,17 +56,17 @@ NULL
 
 
 
-#' @rdname pboostsam
+#' @rdname plagsarlm
 #' @order 1
 #' @export
-pboostsam <- function(
+plagsarlm <- function(
     formula, data = list(), listw, na.action, Durbin, type,
     method = "eigen", quiet = NULL, zero.policy = NULL, interval = NULL,
     tol.solve = .Machine$double.eps, trs = NULL, control = list(),
     stopFun = EBIC, keep = NULL, maxK = NULL, verbose = FALSE) {
 
     # # --- solution A ---
-    # lagsarlm_args <- names(formals(spatialreg::lagsarlm)) # 或 spdep::lagsarlm，取决于你用的包
+    # lagsarlm_args <- names(formals(spatialreg::lagsarlm))
     # user_args <- as.list(environment())
     # args_for_lagsarlm <- user_args[intersect(names(user_args), lagsarlm_args)]
 
@@ -148,4 +153,85 @@ EBIC.Sarlm <- function(object, p, p.keep, ...) {
 
     stopifnot( is.finite(ebic.penalty) )
     BIC(object) + ebic.penalty
+}
+
+
+
+
+
+#' @rdname plagsarlm
+#' @order 2
+#' @export
+plagsarlm2 <- function(x, y, w, maxK = NULL, keep = NULL, verbose = FALSE) {
+    stopifnot( is.matrix(x) )
+    stopifnot( all(rowSums(w) == 1) )
+    stopifnot( !anyNA(x) )
+    stopifnot( !anyNA(y) )
+    stopifnot( !anyNA(w) )
+
+    n <- NROW(x)
+    p <- NCOL(x)
+    ebic.r <- max(0.0, 1.0 - log(n) / (2.0*log(p)))
+    stopifnot( length(y) == n )
+
+    fitsarlm <- function(x, y, w) {
+        stopifnot( NROW(x) == length(y) )
+        stopifnot( is.matrix(x) )
+
+        rho.hat  <- get_rho(x, y, w)
+        A.rho    <- diag(NROW(x)) - rho.hat * w
+        y.tilde  <- A.rho %*% y
+        beta.hat <- coef(lm.fit(x, y.tilde, singular.ok = FALSE))
+        err      <- y.tilde - x %*% beta.hat
+        sig2.hat <- mean( err^2 )
+        BIC      <- n * log(sig2.hat) - 2*log(det(A.rho)) + log(NROW(x)) * sum(NCOL(x))
+        EBIC <- BIC + 2.0 * ebic.r * lchoose(p, NCOL(x))
+
+        list(
+            beta.hat = beta.hat,
+            sig2.hat = sig2.hat,
+            rho.hat = rho.hat,
+            residuals = err,
+            level = EBIC
+        )
+    }
+
+    showiter <- function(verbose, i.max, level=obj.level)
+        if (verbose)
+            if (missing(i.max))
+                message(sprintf("Initial model with level=%.3f", level))
+            else
+                message(sprintf("Adding feature `%i': model level=%.3f", i.max, level))
+
+    if (!is.null(maxK)) {
+        stopifnot(maxK < NCOL(x))
+        maxK <- min( maxK+length(keep), n-1, p-1 )
+    }
+
+    idx <- keep
+    obj <- fitsarlm(x[, idx, drop = FALSE], y, w)
+    obj.level <- obj[["level"]]
+    showiter(verbose, level=obj.level)
+
+    while (TRUE) {
+        i.max <- which.max(abs(crossprod(x, obj[["residuals"]])))
+        idx.tmp <- unique(c(idx, i.max))
+        obj.tmp <- fitsarlm(x[, idx.tmp, drop=FALSE], y, w)
+        obj.tmp.level <- obj.tmp[["level"]]
+        showiter(verbose, i.max, obj.tmp.level)
+
+        if (!is.null(maxK)) {
+            if (length(idx.tmp) > maxK) break
+        } else {
+            if (obj.tmp.level > obj.level) {
+                break
+            }
+        }
+
+        obj <- obj.tmp
+        idx <- idx.tmp
+        obj.level <- obj.tmp.level
+    }
+
+    return(obj)
 }